docs: split test code
